Plutarch.Internal.Ord module, migrate PPartialOrd and POrd
