+++ Plutarch/Integer.hs
@@ -14,10 +14,11 @@ module Plutarch.Integer (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Bool (POrd, PPartialOrd, pif, (#<), (#<=))
+import Plutarch.Bool (pif)
 import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
+import Plutarch.Internal.Ord (POrd, PPartialOrd ((#<), (#<=)))
 import Plutarch.Internal.Other (POpaque, pto)
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (DPTStrat, DerivePlutusType, PInner, PlutusType)
